command系統を抽出しCommandServerとして取り出した
