+++ DodontoFServer.rb
@@ -55,7 +55,7 @@ end
 require "loggingFunction.rb"
 require "FileLock.rb"
 require "saveDirInfo.rb"
-
+require "CommandServer.rb"
 
 $dodontofWarning = nil
 
@@ -102,9 +102,10 @@ $record = 'record.json'
 
 $diceBotTableSaveKey = "diceBotTable"
 
-class DodontoFServer
+class DodontoFServer < CommandServer
   def initialize(saveDirInfo, cgiParams)
-    @cgiParams = cgiParams
+    super cgiParams
+
     @saveDirInfo = saveDirInfo
     
     roomIndexKey = "room"
@@ -135,57 +136,8 @@ class DodontoFServer
       logging(saveFileName, "saveFileName")
       @saveFiles[saveDataKeyName] = @saveDirInfo.getTrueSaveFileName(saveFileName)
     end
-    
   end
 
-  # 次のメソッドをどのモードで追加するか
-  @@next_method_type = :notCommand
-  # CGIに飛んできたコマンドのどれを実際にコマンドとして扱うかのテーブル
-  @@commands = { }
-
-  # コマンド関数を作る
-  def self.command
-    @@next_method_type = :hasReturn
-  end
-
-  # 戻り値を使わないタイプのコマンド関数を作る
-  def self.command_noreturn
-    @@next_method_type = :hasNoReturn
-  end
-
-  # メソッド追加時に呼ばれるコールバック
-  def self.method_added(name)
-    return if @@next_method_type == :notCommand
-    @@commands[name.to_s] = @@next_method_type
-    @@next_method_type = :notCommand
-  end
-
-
-  def getRawCGIValue
-    @cgi ||= CGI.new
-    @cgi.params[key].first
-  end
-  
-  def getRequestData(key)
-    logging(key, "getRequestData key")
-    
-    value = @cgiParams[key]
-    # logging(@cgiParams, "@cgiParams")
-    # logging(value, "getRequestData value")
-    
-    if( value.nil? )
-      if( @isWebIf )
-        value = getRawCGIValue
-      end
-    end
-    
-    
-    # logging(value, "getRequestData result")
-    
-    return value
-  end
-
-  
   attr :isAddMarker
   attr :jsonpCallBack
   attr :isJsonResult
@@ -196,7 +148,6 @@ class DodontoFServer
     return @card unless( @card.nil? )
     
     @card = Card.new();
-    
     
     return @card
   end
@@ -764,29 +715,6 @@ class DodontoFServer
     return empty  if( text.empty? )
     
     return text
-  end
-
-  def analyzeCommand
-    commandName = getRequestData('cmd')
-
-    logging(commandName, "commandName")
-
-    if( commandName.nil? or commandName.empty? )
-      return getResponseTextWhenNoCommandName
-    end
-
-    commandType = @@commands[commandName]
-    logging(commandType, "commandType")
-
-    case commandType
-    when :hasReturn
-      return self.send( commandName )
-    when :hasNoReturn
-      self.send( commandName )
-      return nil
-    else
-      throw Exception.new("\"" + commandName.untaint + "\" is invalid command")
-    end
   end
 
   def getResponseTextWhenNoCommandName
