+++ DodontoFServerMySqlKai.rb
@@ -55,6 +55,7 @@ end
 require "loggingFunction.rb"
 require "FileLock.rb"
 require "saveDirInfoMysql.rb"
+require "CommandServer.rb"
 
 
 $dodontofWarning = nil
@@ -110,10 +111,11 @@ class Mysql::Stmt
   end
 end
 
-class DodontoFServer
+class DodontoFServer < CommandServer
   
   def initialize(saveDirInfo, cgiParams)
-    @cgiParams = cgiParams
+    super cgiParams
+
     @saveDirInfo = saveDirInfo
     
     roomIndexKey = "room"
@@ -138,55 +140,6 @@ class DodontoFServer
     @saveDirInfo.init(roomNumber, $saveDataMaxCount, $SAVE_DATA_DIR)
   end
 
-
-  # 次のメソッドをどのモードで追加するか
-  @@next_method_type = :notCommand
-  # CGIに飛んできたコマンドのどれを実際にコマンドとして扱うかのテーブル
-  @@commands = { }
-
-  # コマンド関数を作る
-  def self.command
-    @@next_method_type = :hasReturn
-  end
-
-  # 戻り値を使わないタイプのコマンド関数を作る
-  def self.command_noreturn
-    @@next_method_type = :hasNoReturn
-  end
-
-  # メソッド追加時に呼ばれるコールバック
-  def self.method_added(name)
-    return if @@next_method_type == :notCommand
-    @@commands[name.to_s] = @@next_method_type
-    @@next_method_type = :notCommand
-  end
-
-
-  def getRawCGIValue
-    @cgi ||= CGI.new
-    @cgi.params[key].first
-  end
-  
-  def getRequestData(key)
-    logging(key, "getRequestData key")
-    
-    value = @cgiParams[key]
-    # logging(@cgiParams, "@cgiParams")
-    # logging(value, "getRequestData value")
-    
-    if( value.nil? )
-      if( @isWebIf )
-        value = getRawCGIValue
-      end
-    end
-    
-    
-    # logging(value, "getRequestData result")
-    
-    return value
-  end
-  
-  
   attr :isAddMarker
   attr :jsonpCallBack
   attr :isJsonResult
@@ -197,7 +150,6 @@ class DodontoFServer
     return @card unless( @card.nil? )
     
     @card = Card.new();
-    
     
     return @card
   end
@@ -444,38 +396,6 @@ class DodontoFServer
     return empty  if( text.empty? )
     
     return text
-  end
-  
-  def analyzeCommand
-    commandName = getRequestData('cmd')
-
-    logging(commandName, "commandName")
-
-    if( commandName.nil? or commandName.empty? )
-      return getResponseTextWhenNoCommandName
-    end
-
-    commandType = @@commands[commandName]
-    logging(commandType, "commandType")
-
-    begin
-      case commandType
-      when :hasReturn
-        return self.send( commandName )
-      when :hasNoReturn
-        self.send( commandName )
-        return nil
-      else
-        throw Exception.new("\"" + commandName.untaint + "\" is invalid command")
-      end
-    ensure
-      begin
-        @db.close unless( @db.nil? )
-      rescue Exception
-        # loggingForce("close Exception")
-        # loggingException(e)
-      end
-    end
   end
 
   def getResponseTextWhenNoCommandName
@@ -6935,7 +6855,16 @@ logging(saveData, 'changePlayRoom() saveData end')
     response = nil
     
     if( $dodontofWarning.nil? )
-      response = analyzeCommand
+      begin
+        response = analyzeCommand
+      ensure
+        begin
+          @db.close unless( @db.nil? )
+        rescue Exception
+          # loggingForce("close Exception")
+          # loggingException(e)
+        end
+      end
     else
       response = {}
       response["warning"] = $dodontofWarning
